REFACTOR : revise the code of controller and configs
